+++ message.h
@@ -19,35 +19,35 @@ static const int g_mask_leading_length = strlen(g_mask_leading_keyword);    ///<
 /**
  * @brief The below definitions are related to the raw data sizes (binary form)
  */
-#define TYPE_SIZE                   UINT8_C(sizeof(uint8_t))
-#define LENGTH_SIZE                 UINT8_C(sizeof(uint8_t))
-#define DATA_SIZE                   UINT8_C(sizeof(uint8_t) * 251)
-#define CRC_SIZE                    UINT8_C(sizeof(uint32_t))
+#define TYPE_SIZE                   (size_t)UINT8_C(sizeof(uint8_t))
+#define LENGTH_SIZE                 (size_t)UINT8_C(sizeof(uint8_t))
+#define DATA_SIZE                   (size_t)UINT8_C(sizeof(uint8_t) * 251)
+#define CRC_SIZE                    (size_t)UINT8_C(sizeof(uint32_t))
 
-#define PAYLOAD_SIZE                UINT8_C(DATA_SIZE + CRC_SIZE)
-#define MASK_SIZE                   UINT8_C(sizeof(uint32_t))
+#define PAYLOAD_SIZE                (size_t)UINT8_C(DATA_SIZE + CRC_SIZE)
+#define MASK_SIZE                   (size_t)UINT8_C(sizeof(uint32_t))
 
-#define ASCII_HEX_LENGTH            UINT8_C(sizeof(uint8_t) * 2)    ///< The length of 1-byte representation in ASCII
+#define ASCII_HEX_LENGTH            (size_t)UINT8_C(sizeof(uint8_t) * 2)    ///< The length of 1-byte representation in ASCII
 
 /**
  * @brief The below definitions are related to the ASCII data sizes representation
  */
-#define TYPE_HEX_LENGTH             UINT8_C(TYPE_SIZE * 2)
-#define LENGTH_HEX_LENGTH           UINT8_C(LENGTH_SIZE * 2)
-#define DATA_HEX_LENGTH             UINT8_C(DATA_SIZE * 2)
-#define CRC32_HEX_LENGTH            UINT8_C(CRC_SIZE * 2)
+#define TYPE_HEX_LENGTH             (size_t)UINT8_C(TYPE_SIZE * 2)
+#define LENGTH_HEX_LENGTH           (size_t)UINT8_C(LENGTH_SIZE * 2)
+#define DATA_HEX_LENGTH             (size_t)UINT8_C(DATA_SIZE * 2)
+#define CRC32_HEX_LENGTH            (size_t)UINT8_C(CRC_SIZE * 2)
 
-#define PAYLOAD_HEX_LENGTH          UINT8_C(PAYLOAD_SIZE * 2)
-#define MASK_HEX_LENGTH             UINT8_C(MASK_SIZE * 2)
+#define PAYLOAD_HEX_LENGTH          (size_t)UINT8_C(PAYLOAD_SIZE * 2)
+#define MASK_HEX_LENGTH             (size_t)UINT8_C(MASK_SIZE * 2)
 
 /**
  * @brief Maximum possible size of a message
  */
-#define ASCII_MESSAGE_MAX_SIZE      (TYPE_HEX_LENGTH + \
+#define ASCII_MESSAGE_MAX_SIZE      (size_t)(TYPE_HEX_LENGTH + \
                                      LENGTH_HEX_LENGTH + \
                                      PAYLOAD_HEX_LENGTH)
 
-#define ASCII_MASK_MAX_SIZE         (MASK_HEX_LENGTH)       ///< The maximum size of the mask in hex ASCII
+#define ASCII_MASK_MAX_SIZE         (size_t)(MASK_HEX_LENGTH)       ///< The maximum size of the mask in hex ASCII
 
 /**
  * @brief Structure that represents the message and its parameters
